% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{add_kegg_subclass_metaboanalyst}
\alias{add_kegg_subclass_metaboanalyst}
\title{get KEGG class and subclass information in order to group pathways into subclass
For MetaboAnalyst results}
\usage{
add_kegg_subclass_metaboanalyst(mA_results, all_kegg_hsa, xeno)
}
\arguments{
\item{mA_results}{metaboAnalyst results}

\item{all_kegg_hsa}{all kegg export}

\item{xeno}{Xenograft of interest}

\item{organism_prefix}{currently set to human "hsa"}
}
\description{
get KEGG class and subclass information in order to group pathways into subclass
For MetaboAnalyst results
}
