% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{add_log2FC}
\alias{add_log2FC}
\title{Add log fold change to the metaboanalyst t-test results}
\usage{
add_log2FC(raw_file, t_res)
}
\arguments{
\item{raw_file}{raw file submitted to metaboanalyst (generated by save_separate_cell_lines)}

\item{metabolome_data}{metaboanalyst t-test result object}
}
\description{
Uses transcriptome sample metadata to subset metabolomics data.
}
