% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{add_overlapping_genes}
\alias{add_overlapping_genes}
\title{Add additional column in the MetaboAnalyst results
Where any of the mapped genes in the DEGs}
\usage{
add_overlapping_genes(
  df,
  deg_universe,
  gene_col = "names",
  new_col = "overlapping_DEGs",
  na_if_empty = TRUE
)
}
\description{
Add additional column in the MetaboAnalyst results
Where any of the mapped genes in the DEGs
}
