% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{conditional_filt}
\alias{conditional_filt}
\title{Function to filter the transcriptome based on occupancy
The occupancy criteria testet are  90\%, 70\%, 50\%}
\usage{
conditional_filt(df, sampledata, comp)
}
\arguments{
\item{df}{Transcriptomics matrix}

\item{sampledata}{sample information dataframe}

\item{comp}{which comparison}
}
\description{
Function to filter the transcriptome based on occupancy
The occupancy criteria testet are  90\%, 70\%, 50\%
}
