% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{count_contrib}
\alias{count_contrib}
\title{Count contribution of genes and metabolites in pathway}
\usage{
count_contrib(df)
}
\arguments{
\item{metaboanalyst_results}{MetaboAnalyst results with all the statistics}

\item{metabolome_data}{MetaboAnalyst mapping file with pathway name and accessions of genes and metabolites}
}
\description{
It takes the metaboAnalyst results and decides contribution of genes and metabolites
contributing the pathways and returns the updated table
}
