% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{create_mixOmics_input}
\alias{create_mixOmics_input}
\title{Separate metabolome and transcriptome data by cell line to be
analysed by mixOmics
Uses transcriptome sample metadata to subset metabolomics data.}
\usage{
create_mixOmics_input(metab_mat, transc_df, meta, comp)
}
\arguments{
\item{metab_mat}{Metabolomics matrix}

\item{transc_df}{Transcriptomics matrix}

\item{meta}{metadata dataframe}

\item{comp}{Comparison of interest}
}
\description{
Separate metabolome and transcriptome data by cell line to be
analysed by mixOmics
Uses transcriptome sample metadata to subset metabolomics data.
}
