% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{extract_link_entities}
\alias{extract_link_entities}
\title{Webgestalt results give a link to view pathways where the metabolites and gene are highlighted
That same link can be used to get metabolite names from ChEBI database
For MetaboAnalyst results}
\usage{
extract_link_entities(link, ref_df, chebi_compounds)
}
\arguments{
\item{link}{webgestalt link}

\item{ref_df}{dataframe}
}
\description{
Webgestalt results give a link to view pathways where the metabolites and gene are highlighted
That same link can be used to get metabolite names from ChEBI database
For MetaboAnalyst results
}
