% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_processing.R
\name{extract_link_entities_lookup}
\alias{extract_link_entities_lookup}
\title{extract link entities lookup
Where any of the mapped genes in the DEGs}
\usage{
extract_link_entities_lookup(link, ref_df, chebi_compounds)
}
\arguments{
\item{link}{link from webgestalt result}

\item{ref_df}{reference dataframe created (rudy_dani_hmdb)}

\item{chebi_compounds}{chebi compounds downloaded from ChEBI}
}
\description{
extract link entities lookup
Where any of the mapped genes in the DEGs
}
