% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{get_hmdb_ids}
\alias{get_hmdb_ids}
\title{Given a list of metabolites short names
pull hmdb ids from annotation dataframe
Annotation dataframe created by Rudy and Daniela}
\usage{
get_hmdb_ids(metab_list, type)
}
\arguments{
\item{metab_list}{metabolite short name list}

\item{type}{whether metabolite list consist of "lipid" or "non-lipid" or
"none" in case we want all of them}
}
\description{
Given a list of metabolites short names
pull hmdb ids from annotation dataframe
Annotation dataframe created by Rudy and Daniela
}
