% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{get_mapping_info}
\alias{get_mapping_info}
\title{Information about the mapping of metabolites/lipids from kit to HMDB
This is to know the percentage of metabolites/lipids having corresponding HMDB ids}
\usage{
get_mapping_info(dataframe)
}
\arguments{
\item{dataframe}{(curated data by Daniela and Rudy)}
}
\description{
Information about the mapping of metabolites/lipids from kit to HMDB
This is to know the percentage of metabolites/lipids having corresponding HMDB ids
}
