% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{getEnrichment}
\alias{getEnrichment}
\title{RaMP DB is not usable for this version of R/bioconductor}
\usage{
getEnrichment(metabolites)
}
\description{
RaMP DB is not usable for this version of R/bioconductor
}
