% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{make_correlation_heatmap}
\alias{make_correlation_heatmap}
\title{Function to make a heatmap of correlation results
to see top 50 correlation pairs}
\usage{
make_correlation_heatmap(cor_dataframe, title)
}
\arguments{
\item{cor_dataframe}{correlation dataframe}

\item{title}{Title}
}
\description{
Function to make a heatmap of correlation results
to see top 50 correlation pairs
}
