% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{make_scatter_plot}
\alias{make_scatter_plot}
\title{Function to make a scatter plot of fold changes of one xenograft vs another xenograft
to see whether upregulation of gene in one xenograft also occurs with upregulation in another xenograft
and vice -vesra}
\usage{
make_scatter_plot(res_df1, res_df2, lab1, lab2)
}
\arguments{
\item{res_df1}{deg results 1}

\item{res_df2}{deg results 2}

\item{lab1}{Label 1}

\item{lab2}{Label 2}
}
\description{
Function to make a scatter plot of fold changes of one xenograft vs another xenograft
to see whether upregulation of gene in one xenograft also occurs with upregulation in another xenograft
and vice -vesra
}
