% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{metaboanalyst_kegg_class}
\alias{metaboanalyst_kegg_class}
\title{get KEGG class and subclass information in order to group pathways into subclass
For MetaboAnalyst results}
\usage{
metaboanalyst_kegg_class(kegg_ids, organism_prefix = "hsa")
}
\arguments{
\item{kegg_ids}{list of KEGG ids}

\item{organism_prefix}{currently set to human "hsa"}
}
\description{
get KEGG class and subclass information in order to group pathways into subclass
For MetaboAnalyst results
}
