% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{normalize_name}
\alias{normalize_name}
\title{Process the results}
\usage{
normalize_name(x)
}
\arguments{
\item{x}{metabolite column name}
}
\description{
As metabolites are column names, R adds X in front of the columns that start with numbers
In order to match metabolite names with annotations we need to handle these cases long with
some other cases
}
