% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{paired_heatmap}
\alias{paired_heatmap}
\title{Function to make a heatmap of correlation results
to see top correlations
and vice -vesra}
\usage{
paired_heatmap(dataFrame, title)
}
\arguments{
\item{dataFrame}{correlation dataframe}

\item{title}{Title}
}
\description{
Function to make a heatmap of correlation results
to see top correlations
and vice -vesra
}
