% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{paired_scatter_plot}
\alias{paired_scatter_plot}
\title{Function to make a scatter plot of correlation results
to see top correlations
and vice -vesra}
\usage{
paired_scatter_plot(dataFrame, title)
}
\arguments{
\item{dataFrame}{correlation dataframe}

\item{title}{Title}
}
\description{
Function to make a scatter plot of correlation results
to see top correlations
and vice -vesra
}
