% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{plot_combined_kegg_pathway}
\alias{plot_combined_kegg_pathway}
\title{Plot combined KEGG pathway
highlighting the genes and metabolites that were queried for pathway enrichment}
\usage{
plot_combined_kegg_pathway(
  pathway_name,
  pathway_id,
  res_map_list,
  species = "hsa",
  file.type = "png",
  out.prefix = NULL
)
}
\arguments{
\item{pathway_name}{Pathway name}

\item{pathway_id}{KEGG pathway accession}

\item{res_map_list}{genes and metabolites to highlight}
}
\description{
Plot combined KEGG pathway
highlighting the genes and metabolites that were queried for pathway enrichment
}
