% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{plot_pathway_presence_heatmap}
\alias{plot_pathway_presence_heatmap}
\title{Count contribution of genes and metabolites in pathway}
\usage{
plot_pathway_presence_heatmap(pathway, res_map_list, title_prefix = NULL)
}
\arguments{
\item{pathway}{pathway of interest}

\item{res_map_list}{Result map list}

\item{title_prefix}{title}
}
\description{
It takes the metaboAnalyst results and decides contribution of genes and metabolites
contributing the pathways and returns the updated table
}
