% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{process_metaboanalyst}
\alias{process_metaboanalyst}
\title{Process metaboAnalyst joint Pathway analysis results}
\usage{
process_metaboanalyst(metaboanalyst_results, metaboanalyst_mapping)
}
\arguments{
\item{metaboanalyst_results}{MetaboAnalyst results with all the statistics}

\item{metabolome_data}{MetaboAnalyst mapping file with pathway name and accessions of genes and metabolites}
}
\description{
It takes gene/metabolite accession and pings KEGG to get common name and
returns the updated table
}
