% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{process_results}
\alias{process_results}
\title{Process the results}
\usage{
process_results(dataframe)
}
\arguments{
\item{dataframe}{raw file submitted to metaboanalyst (generated by save_separate_cell_lines)}
}
\description{
As metabolites are column names, R adds X in front of the columns that start with numbers
In order to match metabolite names with annotations we need to handle these cases long with
some other cases
}
