% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{run_metaboanalyst_ttest}
\alias{run_metaboanalyst_ttest}
\title{Run metaboanalyst t-test}
\usage{
run_metaboanalyst_ttest(input_file, string)
}
\arguments{
\item{input_file}{input file for metaboanalyst}

\item{string}{this string will be prefix to the metafiles generated}
}
\description{
Run metaboanalyst t-test
}
