% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{run_selected_correlation_fdr}
\alias{run_selected_correlation_fdr}
\title{Run correlation analysis with fdr added (FDR is not actually needed here)}
\usage{
run_selected_correlation_fdr(
  transcript_data,
  metabolome_data,
  selected_metabolite
)
}
\arguments{
\item{transcript_data}{filtered/tpm-scaled transcriptome}

\item{metabolome_data}{complete metabolome}

\item{selected_metabolite}{selected metabolites vector (significant results from t-test)}
}
\description{
Uses transcriptome sample metadata to subset metabolomics data.
}
