% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{save_separate_cell_lines}
\alias{save_separate_cell_lines}
\title{Separate metabolome data by cell line}
\usage{
save_separate_cell_lines(transcript_data, metabolome_data, write_file)
}
\arguments{
\item{transcript_data}{Transcriptomics matrix}

\item{metabolome_data}{Metabolomics matrix}

\item{write_file}{Output CSV file}
}
\description{
Uses transcriptome sample metadata to subset metabolomics data.
}
