% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{summarize_corr}
\alias{summarize_corr}
\title{get the summary of correlation results}
\usage{
summarize_corr(
  corr_df,
  top_n = 100,
  p_cutoff = 0.05,
  cor_cutoff = 0.7,
  slope_cutoff = 1
)
}
\arguments{
\item{corr_df}{Correlation results}

\item{top_n}{summary of top n results sorted by correlation coefficient}

\item{p_cutoff}{p-value cutoff currently set to 0.05}

\item{cor_cutoff}{cor_cutoff cutoff currently set to 0.7 positive or negative}

\item{slope_cutoff}{slope_cutoff cutoff currently set to 1 positive or negative}
}
\description{
get the summary of correlation results
}
