% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{webgestalt_kegg_class}
\alias{webgestalt_kegg_class}
\title{get KEGG class and subclass information in order to group pathways into subclass
For webgestalt results}
\usage{
webgestalt_kegg_class(webgestalt, organism_prefix = "hsa")
}
\arguments{
\item{webgestalt}{webgestalt results}

\item{organism_prefix}{currently set to human "hsa"}
}
\description{
get KEGG class and subclass information in order to group pathways into subclass
For webgestalt results
}
